% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{biais_pi}
\alias{biais_pi}
\title{calcule le biais de la probabilite estimee de toxicite selon une Bernoulli(p)
Compare avec la valeur theorique.}
\usage{
biais_pi(n, p)
}
\arguments{
\item{n}{taille de l'echantillon}

\item{p}{parametre de la loi de Bernoulli utilisee}
}
\value{
valeur du biais retourné
}
\description{
calcule le biais de la probabilite estimee de toxicite selon une Bernoulli(p)
Compare avec la valeur theorique.
}
\examples{
test_biais<-biais_pi(18,0.33)
}
