% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{fonction_biais_survie}
\alias{fonction_biais_survie}
\title{Biais.}
\usage{
fonction_biais_survie(n, lambda, t_star)
}
\arguments{
\item{n}{taille d'échantillon}

\item{lambda}{paramètre d'échelle de la loi de weibull}

\item{t_star}{fin de la fenêtre d'observation}
}
\value{
Calcul du biais de la probabilite de toxicite estimee par la fonction simul_survie.
}
\description{
Biais.
}
