% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{fonction_biais_survie_weibull}
\alias{fonction_biais_survie_weibull}
\title{Calculer le biais d'une Weibull k,lambda.}
\usage{
fonction_biais_survie_weibull(n, lambda, k, t_star)
}
\arguments{
\item{t_star}{}
}
\description{
Calculer le biais d'une Weibull k,lambda.
}
\examples{
N<-20
lambda_test<-1/3
k<-3
n<-100
t_star<-6
test_biais_weibull<-fonction_biais_survie_weibull(n,lambda=lambda_test,k,t_star)
}
