% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_diffs.R
\name{fonction_compar_plots}
\alias{fonction_compar_plots}
\title{Plot des valeurs des biais moyens selon la taille des echantillons.}
\usage{
fonction_compar_plots(limit_inf, limit_sup, N, p, lambda, t_star, K, sh)
}
\arguments{
\item{limit_inf}{taille minimale des echantillons}

\item{limit_sup}{taille maximale des echantillons}

\item{N}{nombre de tailles d'echantillon differents.}

\item{p}{probabilite selon le modele de guerison de toxicite.}

\item{lambda}{parametre d'échelle (scale) de la loi weibull.}

\item{t_star}{fin de la fenetre d'observation}

\item{K}{nombre de simulations de la meme taille d'echantillon.}

\item{sh}{paramètre de la loi weibull.}
}
\value{
Plot des valeurs des biais moyens selon la taille des echantillons.
}
\description{
Plot des valeurs des biais moyens selon la taille des echantillons.
}
\examples{
######Test ######
p2<-0.3
k<-50
lambda7<-0.33
t_star<-6
lmoins<-1
l_plus<-1000
N<-50
shape<-2
test_plot<-fonction_compar_plots(limit_sup = l_plus,limit_inf = lmoins,
N=N,p=p2,lambda=lambda7,t_star=t_star,K=k,sh=shape)
}
