% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_diffs.R
\name{fonction_compar_plotsn_lambda}
\alias{fonction_compar_plotsn_lambda}
\title{Plot des valeurs des biais moyens selon la taille des echantillons et du lambda.}
\usage{
fonction_compar_plotsn_lambda(N, window_lambda, t_star)
}
\arguments{
\item{N}{nombre de tailles d'echantillon differents.}

\item{window_lambda}{vecteur de valeurs pour lambda}

\item{t_star}{fin de la fenetre d'observation}
}
\value{
Plot des valeurs des biais moyens en fonction du lambda et de la taille des echantillons.
}
\description{
Plot des valeurs des biais moyens selon la taille des echantillons et du lambda.
}
\examples{
######Test ######
t_star<-3
N<-10
window_lambda<-c(0.2,0.5,0.1)
result<-fonction_compar_plotsn_lambda(N,window_lambda,t_star)
}
