% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_taille_differente.R
\name{fonction_generation_taille_differente}
\alias{fonction_generation_taille_differente}
\title{Permet de renvoyer le biais pour des tailles differentes.}
\usage{
fonction_generation_taille_differente(vector_size, model, liste_parameter)
}
\arguments{
\item{vector_size}{vecteur des différentes tailles.}

\item{model}{renverra a surv ou a bern. Correspondra aux simulations differentes.}

\item{liste_parameter}{Correspondra a p(bernoulli) ou a lambda et t_star pour exp.}
}
\value{
Vecteur des biais pour chaque taille.
}
\description{
Permet de renvoyer le biais pour des tailles differentes.
}
\examples{
#################TEST EXP #####################
vecteur_size<-sample(c(1:100),10)
lamdba_test<-0.33
t_star<-6
liste_parameter<-list(lambda_test,t_star)
names(liste_parameter)<-c("lambda","t_star")
modele<-"surv"
test_generation_taillediff_exp<-fonction_generation_taille_differente(vector_size=vecteur_size,model=modele,liste_parameter = liste_parameter)

#################TEST BERNOULLI ###################
prop<-0.33
list_param<-list(prop)
names(list_param)<-c("p")
modele2<-"bernoulli"
test_generation_taillediff_bern<-fonction_generation_taille_differente(vector_size = vecteur_size,model=modele2,liste_parameter = list_param)
}
