% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_mean.R
\name{fonction_generation_taille_mean}
\alias{fonction_generation_taille_mean}
\title{Génére n fois le modèle}
\usage{
fonction_generation_taille_mean(vector_size, liste_parameter, K)
}
\arguments{
\item{vector_size}{: un vecteur de taille d'échantillon}

\item{liste_parameter}{: liste des paramètres}

\item{K}{: nombre de fois que l'on génère un échantillon de taille n}
}
\value{
: Génére n fois le modèle (bernouilli ou survie)
}
\description{
Génére n fois le modèle
}
\examples{
#test avec le modèle "survie"
#---------------------------
N<-10
vecteur_size<-sample(c(1:100),N)
lambda_test<-3
t_star<-6
p<-0.33
k<-1
liste_parameter<-list(lambda_test,t_star,p=p,k)
names(liste_parameter)<-c("lambda","t_star","p","k")
K2<-20
test_exp_taillemoy<-fonction_generation_taille_mean(vector_size=vecteur_size,K=K2,liste_parameter = liste_parameter)
#################### Plot des r?sultats en fonction de la taille.########
donnees_taille_biaismoyen<-cbind.data.frame(vecteur_size[order(vecteur_size)],test_exp_taillemoy)
colnames(donnees_taille_biaismoyen)<-c("Size","Mean_Bias_Cure","Mean_Bias_Surv")
plot(donnees_taille_biaismoyen$Size,donnees_taille_biaismoyen$Mean_Bias_Cure,main="The mean bias according to the size with Survival function")
points(x=donnees_taille_biaismoyen$Size,y=donnees_taille_biaismoyen$Mean_Bias_Surv,col="red")
}
