% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_taille_differente.R
\name{fonction_graph_fonc_size}
\alias{fonction_graph_fonc_size}
\title{Renvoie la valeur du biais et la taille des echantillons dans la meme table.}
\usage{
fonction_graph_fonc_size(vector_size, model, liste_parameter)
}
\arguments{
\item{vector_size}{vecteur correspondant aux differentes tailles d'echantillons.}

\item{model}{chaine de caracteres.}

\item{liste_parameter}{: liste de parametres. Vaudra p (pour Bernoulli) ou lambda et t_star pour survie}
}
\value{
Renvoie la valeur du biais et la taille des echantillons dans la meme table.
}
\description{
Renvoie la valeur du biais et la taille des echantillons dans la meme table.
}
\examples{
######1) exp#####
N<-50
vecteur_size<-sample(c(10:1000),N)
lamdba_test<-0.33
t_star<-6
liste_parameter<-list(lambda_test,t_star)
names(liste_parameter)<-c("lambda","t_star")
modele<-"surv"
test_graph_exp<-fonction_graph_fonc_size(vector_size = vecteur_size,model=modele,liste_parameter=liste_parameter)
}
