% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{function_influence_rate}
\alias{function_influence_rate}
\title{Generer le biais en fonction d'un nombre de k (shape) differents.}
\usage{
function_influence_rate(n, lim_moins, lim_plus, lambda, t_star, number_k)
}
\arguments{
\item{n}{taille de l'echantillon.}

\item{lim_moins}{valeur minimale de la fenetre des k.}

\item{lim_plus}{valeur maximale de la fenetre des k.}

\item{lambda}{valeur de la loi Weibull (scale).}

\item{t_star}{fin de la fenetre d'observation.}

\item{number_k}{Nombre de k differents etudies.}
}
\value{
Valeur du biais pour tous les k differents etudies.
}
\description{
Generer le biais en fonction d'un nombre de k (shape) differents.
}
\examples{
number_trials<-10
l_plus<-5
l_moins<-0.1
vecteur_k_bias<-function_influence_rate(n,lim_moin=l_moins,lim_plus=l_plus,
lambda=lambda_test,number_k=number_trials,t_star=t_star)
}
