% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_func.R
\name{lambda_func}
\alias{lambda_func}
\title{Calcule le biais à partir d'un vecteur de lambdas}
\usage{
lambda_func(lambda_vec, list_params)
}
\arguments{
\item{lambda_vec}{: vecteur de lambdas}

\item{list_params}{: liste de paramètres nécessaires (ici n : taille de l'échantillon, t_star : fenêtre d'obs)}
}
\value{
retourne un vecteur de biais
}
\description{
Calcule le biais à partir d'un vecteur de lambdas
}
\examples{
n<-100
t_star<-6
lambda_vec <- c(0.8,0.41, 0.28, 0.25,0.42, 0.04)
liste_parameter<-list(n,t_star)
names(liste_parameter)<-c("n","t_star")
test <-lambda_func(lambda_vec=lambda_vec,list_params =  liste_parameter)

#plot du résultat
#----------------
ggplot(data.frame(x = lambda_vec, y = test), aes(x = x, y = y)) + geom_line(color = "blue", size = 1.5) + ggtitle("Bias with different lambdas") + xlab("lambda") + ylab("Biais") + theme_minimal()
}
