% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_diffs.R
\name{NSimulations.selon.n}
\alias{NSimulations.selon.n}
\title{Matrice composee des biais moyens associes a la taille de l'echantillon de n=20 a n=200 par saut de 20.}
\usage{
NSimulations.selon.n(N, lambda, t_star)
}
\arguments{
\item{N}{nombre de tailles d'echantillon differents.}

\item{lambda}{parametre de la loi exponentielle.}

\item{t_star}{fin de la fenetre d'observation}
}
\value{
Valeur du biais moyen selon n dans l'intervalle (20,200).
}
\description{
Matrice composee des biais moyens associes a la taille de l'echantillon de n=20 a n=200 par saut de 20.
}
\examples{
######Test ######
t_star<-3
N<-10
lambda<-c(0.2)
result<-NSimulations.selon.n(N,lambda,t_star)
}
