% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_func.R
\name{pbinom_func}
\alias{pbinom_func}
\title{Génère le biais à partir d'un vecteur de probabilité p}
\usage{
pbinom_func(p_vec, n)
}
\arguments{
\item{p_vec}{: vecteur de probabilités}

\item{n}{: taille de l'échantillon}
}
\value{
retourne un vecteur de biais
}
\description{
Génère le biais à partir d'un vecteur de probabilité p
}
\examples{
vec_p <- c(0.2,0.33,0.24,0.12,0.18)
test1 <- pbinom_func(vec_p, n)

#plot du résultat
#----------------
ggplot(data.frame(x = vec_p, y = test1), aes(x = x, y = y)) + geom_line(color = "blue", size = 1.5) + ggtitle("Bias with different probabilities") + xlab("p") + ylab("Bias") + theme_minimal()
}
