% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{simul_bernoulli}
\alias{simul_bernoulli}
\title{simule un n_echantillon d'une Bernoulli(p).}
\usage{
simul_bernoulli(n, p)
}
\arguments{
\item{n}{taille de l'echantillon}

\item{p}{parametre de la loi de Bernoulli}
}
\value{
un vecteur de taille n representant un n-echantillon de Bernoulli
}
\description{
simule un n_echantillon d'une Bernoulli(p).
}
\examples{
n<-10
test<-simul_bernoulli(100,0.1)
test
}
