% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{simul_exp}
\alias{simul_exp}
\title{Simuler un échantillon de taille n selon la loi exponentiel de paramètre lambda.}
\usage{
simul_exp(n, lambda)
}
\arguments{
\item{n}{: taille de l'échantillon}

\item{lambda}{: paramètre de l'exponentiel.}
}
\value{
Un échantillon de taille n.
}
\description{
Simuler un échantillon de taille n selon la loi exponentiel de paramètre lambda.
}
\examples{
simul_exp(10,0.3)
}
