% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{simul_survie}
\alias{simul_survie}
\title{Retourner la probabilité de manifester la toxicité de 0 à t_star.}
\usage{
simul_survie(n, lambda, t_star)
}
\arguments{
\item{n}{: taille de l'échantillon}

\item{lambda}{: paramètre de la loi exponentielle.}

\item{t_star}{: int, fin de la fenêtre d'observation.}
}
\value{
le quantile de la fonction de survie en t_star.
}
\description{
Retourner la probabilité de manifester la toxicité de 0 à t_star.
}
\examples{
set.seed(133)
n<-100
test_surv<-simul_survie(n,0.5,6)
}
