% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{simul_survie_weibull}
\alias{simul_survie_weibull}
\title{Trouver le quantile a t_star de la fonction de survie (Kaplan-Meier).}
\usage{
simul_survie_weibull(n, lambda, k, t_star)
}
\arguments{
\item{n}{Taille de l'echantillon.}

\item{lambda}{parametre de la fonction Weibull (scale).}

\item{k}{parametre de la fonction Weibull (shape).}

\item{t_star}{fin de la fenetre d'observation.}
}
\value{
Trouver le quantile a t_star de la fonction de survie (Kaplan-Meier).
}
\description{
Trouver le quantile a t_star de la fonction de survie (Kaplan-Meier).
}
\examples{
lambda_test<-1/3
k<-3
n<-100
t_star<-6
}
