% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{simul_weibull}
\alias{simul_weibull}
\title{Renvoie une simulation d'une Weibull de parametres k et lambda.}
\usage{
simul_weibull(n, lambda, k)
}
\arguments{
\item{n}{Nombre d'individus a generer.}

\item{lambda}{parametre correspondant a l'echelle (scale) de la loi Weibull generee.}

\item{k}{parametre correspondant a la forme (shape) de la loi Weibull generee.}
}
\value{
Renvoie une simulation d'une Weibull de parametres k et lambda.
}
\description{
Renvoie une simulation d'une Weibull de parametres k et lambda.
}
\examples{
test_simul<-simul_weibull(100,0.1,1)
}
