% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{Simuler_Nfois_n_echantillons}
\alias{Simuler_Nfois_n_echantillons}
\title{Generer N fois un echantillon de taille n}
\usage{
Simuler_Nfois_n_echantillons(N, n, lambda, t_star)
}
\arguments{
\item{N}{nombre de fois generation echantillon}

\item{n}{taille echantillonn}

\item{lambda}{parametre loi exponentielle.}

\item{t_star}{fin de la fenetre.}
}
\value{
Generer N fois un echantillon de taille n
}
\description{
Generer N fois un echantillon de taille n
}
\examples{
lambda_test<-1/3
n<-100
t_star<-6
N<-100
test_simul_total<-Simuler_Nfois_n_echantillons(N,n,lambda_test,t_star)
}
