% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{Simuler_Nfois_n_echantillons_bern}
\alias{Simuler_Nfois_n_echantillons_bern}
\title{retourne un vecteur de taille N constitue des biais issu d'une simulation de
N echantillons de taille n pour une loi de Bernoulli}
\usage{
Simuler_Nfois_n_echantillons_bern(N, n, p)
}
\arguments{
\item{N}{le nombre d'echantillons generes}

\item{n}{la taille de chacun des N echantillons generes}

\item{p}{parametre de la loi de Bernoulli}
}
\value{
un vecteur de taille N
}
\description{
retourne un vecteur de taille N constitue des biais issu d'une simulation de
N echantillons de taille n pour une loi de Bernoulli
}
\examples{
n<-10
N<-100
p<-0.33
test_simul_bern_total<-Simuler_Nfois_n_echantillons_bern(N,n,p)
}
