% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{Simuler_Nfois_n_weibull}
\alias{Simuler_Nfois_n_weibull}
\title{Generer N fois un echantillon de taille n et trouver le biais.}
\usage{
Simuler_Nfois_n_weibull(N, n, lambda, k, t_star)
}
\arguments{
\item{N}{nombre de simulation}

\item{n}{taille de l'echantillon}

\item{lambda}{parametre de la Weibull, scale.}

\item{k}{parametre de la weibull, shape.}

\item{t_star}{fin de la fenetre d'observation.}
}
\value{
valeur du biais pour les N simulations.
}
\description{
Generer N fois un echantillon de taille n et trouver le biais.
}
\examples{
lambda_test<-1/3
k<-3
n<-100
t_star<-6
vecteur<-Simuler_Nfois_n_weibull(N,n,lambda=lambda_test,k,t_star)
}
