% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_func.R
\name{pbinom_func}
\alias{pbinom_func}
\title{Génère le biais à partir d'un vecteur de probabilité p}
\usage{
pbinom_func(p_vec, n, N)
}
\arguments{
\item{p_vec}{: vecteur de probabilités}

\item{n}{: taille de l'échantillon}

\item{N}{Nombre d'échantillon}
}
\value{
retourne un vecteur de biais
}
\description{
Génère le biais à partir d'un vecteur de probabilité p
}
\examples{
vec_p <- c(0.2,0.33,0.24,0.12,0.18)
n <- 100
N<-20
test1 <- pbinom_func(vec_p, n,N=N)

#plot du résultat
#----------------
ggplot2::ggplot(data.frame(x = vec_p, y = test1),
ggplot2::aes(x = x, y = y)) +
ggplot2::geom_line(color = "blue", size = 1.5) +
ggplot2::ggtitle("Bias with different probabilities") +
ggplot2::xlab("p") +
ggplot2::ylab("Bias") +
ggplot2::theme_minimal()
}
